+++ src/Pdf/PdfPreview.php
@@ -8,8 +8,11 @@
 
 namespace HeimrichHannot\UtilsBundle\Pdf;
 
+use Ghostscript\Transcoder;
 use HeimrichHannot\UtilsBundle\Cache\FileCache;
+use HeimrichHannot\UtilsBundle\Container\ContainerUtil;
 use Spatie\PdfToImage\Pdf;
+use Symfony\Component\Config\Definition\Exception\InvalidTypeException;
 
 class PdfPreview
 {
@@ -21,12 +24,17 @@ class PdfPreview
      * @var string
      */
     private $webDir;
+    /**
+     * @var ContainerUtil
+     */
+    private $containerUtil;
 
-    public function __construct(FileCache $cache, string $webDir)
+    public function __construct(FileCache $cache, ContainerUtil $containerUtil, string $webDir)
     {
         $this->cache = $cache;
         $this->cache->setNamespace('pdfPreview');
         $this->webDir = $webDir.'/..';
+        $this->containerUtil = $containerUtil;
     }
 
     /**
@@ -49,24 +57,62 @@ class PdfPreview
     /**
      * Generate a image preview of the given pdf.
      *
-     * Possible file extension: jpg, jpeg, png
+     * Possible PdfTranscoder: spatie (spatie/pdf-to-image), alchemy (alchemy/ghostscript)
+     *
+     * Possible file extensions: jpg, jpeg, png
      *
      * Additional options:
+     * string pdfTranscoder The pdf transcoder to use (default: spatie)
      * int page The page to render (default: 1)
-     * int compressionQuality Pdf compression quality (default: null)
-     * int resolution Raster resolution (default: 144)
+     * int compressionQuality Pdf compression quality (default: null) (spatie only)
+     * int resolution Raster resolution (default: 144)(spatie only)
+     * bool absolutePdfPath Set true if pdf path is absolute (default: false)
+     * bool absoluteImagePath Set true if image path is absolute (default: false)
      *
-     * @param string $pdfPath   the path to the pdf file
-     * @param string $imagePath the path where the image file should be saved (including file name and extension)
+     * @param string $pdfPath   the relative path to the pdf file
+     * @param string $imagePath the relative path where the image file should be saved (including file name and extension)
      * @param array  $options   Additional rendering options
+     *
+     * @throws \Exception
      *
      * @return bool
      */
     public function generatePdfPreview(string $pdfPath, string $imagePath, array $options = [])
     {
+        if (!isset($options['absolutePdfPath']) || true !== $options['absolutePdfPath']) {
+            $pdfPath = $this->webDir.'/'.$pdfPath;
+        }
+        if (!isset($options['absoluteImagePath']) || true !== $options['absoluteImagePath']) {
+            $imagePath = $this->webDir.'/'.$imagePath;
+        }
+        $pdfTranscoder = isset($options['pdfTranscoder']) ? $options['pdfTranscoder'] : '';
+        switch ($pdfTranscoder) {
+            case 'alchemy':
+                return $this->alchemyPdf($pdfPath, $imagePath, $options);
+            case 'spatie':
+            default:
+                return $this->spatiePdf($pdfPath, $imagePath, $options);
+        }
+    }
+
+    /**
+     * @param string $pdfPath
+     * @param string $imagePath
+     *
+     * @throws \Exception
+     *
+     * @return bool
+     */
+    protected function spatiePdf(string $pdfPath, string $imagePath, array $options = [])
+    {
+        try {
+            $this->containerUtil->isBundleActive('spatie/pdf-to-image');
+        } catch (\Exception $e) {
+            throw new \Exception('Package spatie/pdf-to-image is not installed. Please install or use another pdf ttranscoder.');
+        }
         $imageExtension = pathinfo($imagePath, PATHINFO_EXTENSION);
         try {
-            $pdf = new Pdf($this->webDir.'/'.$pdfPath);
+            $pdf = new Pdf($pdfPath);
             if (isset($option['page']) && $options['page'] > 0) {
                 $pdf->setPage($options['page']);
             }
@@ -79,7 +125,54 @@ class PdfPreview
             if (!empty($imageExtension)) {
                 $pdf->setOutputFormat($imageExtension);
             }
-            $pdf->saveImage($this->webDir.'/'.$imagePath);
+            $pdf->saveImage($imagePath);
+        } catch (\Exception $e) {
+            return false;
+        }
+
+        return true;
+    }
+
+    /**
+     * @param string $pdfPath
+     * @param string $imagePath
+     *
+     * @throws \Exception
+     *
+     * @return bool
+     */
+    protected function alchemyPdf(string $pdfPath, string $imagePath, array $options = [])
+    {
+        try {
+            $this->containerUtil->isBundleActive('alchemy/ghostscript');
+        } catch (\Exception $e) {
+            throw new \Exception('Package alchemy/ghostscript is not installed. Please install or use another pdf ttranscoder.');
+        }
+        $imageExtension = pathinfo($imagePath, PATHINFO_EXTENSION);
+        $allowedExtensions = ['jpg', 'jpeg', 'png'];
+        if (!in_array($imageExtension, $allowedExtensions, true)) {
+            throw new InvalidTypeException('Only one of the following file types is allowed: '.implode(
+                ', ', $allowedExtensions)
+            );
+        }
+        if ('jpg' === $imageExtension) {
+            $imageExtension = 'jpeg';
+        }
+        $command = [
+            '-sDEVICE='.$imageExtension,
+            '-dNOPAUSE',
+            '-dBATCH',
+            '-dSAFER',
+            '-sOutputFile='.$imagePath,
+        ];
+        if (isset($option['page']) && is_int($options['page']) && $options['page'] > 0) {
+            $command[] = sprintf('-dFirstPage=%d', $options['page']);
+            $command[] = sprintf('-dLastPage=%d', $options['page']);
+        }
+        try {
+            $command[] = $pdfPath;
+            $transcoder = Transcoder::create();
+            $transcoder->command($command);
         } catch (\Exception $e) {
             return false;
         }